re code the all analyzer
