+++ DevelopmentTools/ngHBAnalyzer.cc
@@ -1,0 +1,1 @@
+#include "ngHBAnalyzer.h"
